requirements.txt
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
psycopg2-binary==2.9.10
